+++ dns_exporter/dns_exporter.py
@@ -1,4 +1,5 @@
 """dns_exporter is a blackbox-style Prometheus exporter for DNS."""
+import argparse
 import ipaddress
 import logging
 import random
@@ -33,30 +34,26 @@ from prometheus_client import (
 )
 from prometheus_client.registry import RestrictedRegistry
 
+# get version number from package metadata if possible
 __version__: str = "0.0.0"
 try:
     __version__ = version("dns_exporter")
 except PackageNotFoundError:
-    # package is not installed
+    # package is not installed, get version from file
     try:
         from _version import version as __version__  # type: ignore
     except ImportError:
+        # this must be a git checkout with no _version.py file, version unknown
         pass
 
+# initialise logger
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 ch = logging.StreamHandler()
 ch.setLevel(logging.DEBUG)
-
-# create formatter
 formatter = logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(message)s")
-
-# add formatter to ch
 ch.setFormatter(formatter)
-
-# add ch to logger
 logger.addHandler(ch)
-
 logger.info(f"dns_exporter v{__version__} starting up")
 
 # create seperate registry for the dns query metrics
@@ -94,7 +91,6 @@ QUERY_FAILURE = Enum(
     "dns_query_failure_reason",
     "The reason this DNS query failed",
     states=[
-        "initial_state",
         "invalid_request_module",
         "invalid_request_target",
         "invalid_request_family",
@@ -180,15 +176,13 @@ class DNSRequestHandler(MetricsHandler):
     ) -> bool:
         """Validate the incoming scrape HTTP request before doing the DNS query."""
         # do we have a module in the request?
-        if "module" not in query or query["module"] not in config["modules"]:  # type: ignore
+        if "module" in query and ("modules" not in config or query["module"] not in config["modules"]):  # type: ignore
+            logger.warning(
+                "Scrape request contains a module '{query['module']}' but the module is unknown to this exporter."
+            )
             QUERY_SUCCESS.set(0)
             QUERY_FAILURE.state("invalid_request_module")
             return False
-        elif "target" not in query:
-            QUERY_SUCCESS.set(0)
-            QUERY_FAILURE.state("invalid_request_target")
-            return False
-
         # all good
         return True
 
@@ -217,9 +211,10 @@ class DNSRequestHandler(MetricsHandler):
             "validate_additional_rrs": {},
         }
 
-        # the module config has middle precedence
-        assert hasattr(self, "config")  # mypy
-        self.module.update(self.config["modules"][qs["module"]])
+        # the module from the config file has middle precedence,
+        # if one was specified in the querystring
+        if "module" in qs and hasattr(self, "config") and "modules" in self.config:
+            self.module.update(self.config["modules"][qs["module"]])
 
         # and the querystring from the scrape request has highest precedence,
         # overruling values from defaults and module
@@ -625,6 +620,7 @@ class DNSRequestHandler(MetricsHandler):
                 self.send_metric_response(registry=dns_registry, query=self.qs)
                 return
 
+            logger.debug(f"Got scrape request from client {self.client_address}")
             # assemble module (configuration for this scrape) from defaults, config file and request
             self.get_module(qs=self.qs)
             if not self.validate_module():
@@ -691,9 +687,11 @@ class DNSRequestHandler(MetricsHandler):
                     query=q,
                     timeout=float(str(self.module["timeout"])),
                 )
+                logger.debug("Got a DNS query response!")
             except dns.exception.Timeout:
                 # configured timeout was reached before we got a response
                 QUERY_FAILURE.state("timeout")
+                logger.error("DNS query timeout was reached")
             except Exception:
                 logger.exception(
                     f"Got an exception while module {self.qs['module']} was looking up qname {self.module['query_name']} using target {self.module['target']}"
@@ -708,6 +706,7 @@ class DNSRequestHandler(MetricsHandler):
                 DNS_FAILURES.inc()
                 QUERY_SUCCESS.set(0)
                 self.send_metric_response(registry=dns_registry, query=self.qs)
+                logger.debug("Returning DNS query metrics - no response received :(")
                 return
 
             # we got a response, increase the response counter
@@ -763,11 +762,13 @@ class DNSRequestHandler(MetricsHandler):
             self.send_metric_response(
                 registry=dns_registry, query=self.qs, skip_metrics=skip_metrics
             )
+            logger.debug(f"Returning DNS query metrics - query success: {success}")
             return
 
         # this endpoint exposes metrics about the exporter itself and the python process
         elif self.url.path == "/metrics":
             self.send_metric_response(registry=self.registry, query=self.qs)
+            logger.debug("Returning exporter metrics for request to /metrics")
             return
 
         # the root just returns a bit of informational html
@@ -777,25 +778,116 @@ class DNSRequestHandler(MetricsHandler):
             self.end_headers()
             self.wfile.write(INDEX.encode("utf-8"))
             HTTP_RESPONSES.labels(path="/", response_code=200).inc()
+            logger.debug("Returning index page for request to /")
 
         # unknown endpoint
         else:
             self.send_response(404)
             self.end_headers()
             HTTP_RESPONSES.labels(path=self.url.path, response_code=404).inc()
+            logger.debug(f"Unknown endpoint '{self.url.path}' returning 404")
             return None
 
 
-def main() -> None:
+def get_parser() -> argparse.ArgumentParser:
+    """Create and return the argparse object."""
+    parser = argparse.ArgumentParser(
+        description=f"dns_exporter version {__version__}. See ReadTheDocs for more info."
+    )
+
+    # optional arguments
+    parser.add_argument(
+        "-c",
+        "--config-file",
+        dest="config-file",
+        help="The path to the yaml config file to use. Only the root 'modules' key is read from the config file.",
+        default=argparse.SUPPRESS,
+    )
+    parser.add_argument(
+        "-d",
+        "--debug",
+        action="store_const",
+        dest="log-level",
+        const="DEBUG",
+        help="Debug mode. Equal to setting --log-level=DEBUG.",
+        default=argparse.SUPPRESS,
+    )
+    parser.add_argument(
+        "-l",
+        "--log-level",
+        dest="log-level",
+        choices=["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"],
+        help="Logging level. One of DEBUG, INFO, WARNING, ERROR, CRITICAL. Defaults to INFO.",
+        default=argparse.SUPPRESS,
+    )
+    parser.add_argument(
+        "-q",
+        "--quiet",
+        action="store_const",
+        dest="log-level",
+        const="WARNING",
+        help="Quiet mode. No output at all if there is nothing to do, and no errors are encountered. Equal to setting --log-level=WARNING.",
+        default=argparse.SUPPRESS,
+    )
+    parser.add_argument(
+        "-v",
+        "--version",
+        dest="version",
+        action="store_true",
+        help="Show version and exit.",
+        default=argparse.SUPPRESS,
+    )
+    return parser
+
+
+def parse_args(
+    mockargs: Optional[list[str]] = None,
+) -> tuple[argparse.ArgumentParser, argparse.Namespace]:
+    """Create an argparse monster and parse mockargs or sys.argv[1:]."""
+    parser = get_parser()
+    args = parser.parse_args(mockargs if mockargs else sys.argv[1:])
+    return parser, args
+
+
+def main(mockargs: Optional[list[str]] = None) -> None:
     """Read config and start exporter."""
-    with open("dns_exporter.yml") as f:
-        try:
-            config = yaml.load(f, Loader=yaml.SafeLoader)
-        except Exception:
-            logger.exception(
-                "Unable to parse YAML config file dns_exporter.yml - bailing out."
+    # get arpparser and parse args
+    parser, args = parse_args(mockargs)
+
+    # handle a couple of special cases before reading config
+    if hasattr(args, "version"):
+        print(f"dns_exporter version {__version__}")
+        sys.exit(0)
+
+    if hasattr(args, "config-file"):
+        with open(getattr(args, "config-file"), "r") as f:
+            try:
+                config = yaml.load(f, Loader=yaml.SafeLoader)
+            except Exception:
+                logger.exception(
+                    f"Unable to parse YAML config file {getattr(args, 'config-file')} - bailing out."
+                )
+                sys.exit(1)
+        if (
+            not config
+            or "modules" not in config
+            or not isinstance(config["modules"], dict)
+            or not config["modules"]
+        ):
+            # config is empty, missing "modules" key, or modules is empty or not a dict
+            logger.error(
+                f"Invalid config file {getattr(args, 'config-file')} - yaml was valid but no modules found"
             )
             sys.exit(1)
+        logger.debug(
+            f"The following modules were loaded from config file {getattr(args, 'config-file')}: {list(config['modules'].keys())}"
+        )
+    else:
+        # we have no config file
+        config = {}
+        logger.debug("No -c / --config-file found so a config file will not be used.")
+
+    # initialise handler and start HTTPServer
     handler = DNSRequestHandler
     handler.config = config  # type: ignore
     HTTPServer(("127.0.0.1", 15353), handler).serve_forever()