+++ CHANGELOG.md
@@ -5,9 +5,18 @@ All notable changes to `dns_exporter` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v0.2.0] - 2022-01-08
 
-This is the initial release.
+## [v0.2.0-beta3] - 2022-01-09
+
+### Changed
+- Removed the config file requirement
+- Removed the requirement to supply a module in every scrape request
+- Updated README.md with more information
+
+
+## [v0.2.0-beta2] - 2022-01-08
+
+This was the first public pre-release.
 
 ### Added